[Protocol] Refactoring requests and responses into their own folders
