+++ src/Gnugat/SoulMeMaybe/NetSoulProtocol/Request/AbstractRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Gnugat\SoulMeMaybe\NetSoulProtocol;
+namespace Gnugat\SoulMeMaybe\NetSoulProtocol\Request;
 
 /**
  * Abstract request class.
