+++ src/Gnugat/SoulMeMaybe/NetSoulProtocol/Response/AbstractResponse.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Gnugat\SoulMeMaybe\NetSoulProtocol;
+namespace Gnugat\SoulMeMaybe\NetSoulProtocol\Response;
 
 /**
  * Abstract response class.
